Converted FullTextSessionFactory into an interface and added an implementation
